Update obj-c tests to remove uses of deprecated methods
A bunch of tests are now redundant or meaningless and were removed. The
remaining tests were reorganized and improved.

Everything related to RLMRealmConfiguration validation has been pushed into
tests for that and out from the realm and migration tests.

The remaining Realm tests have been reorganized to put them in logical groups
with section markers.

The migration tests have been split into separate tests for if a migration is
needed, if the migration is called at the correct time, and if the actual
migration works correctly. When exactly a schema version bump is needed should
now be fully specified by the tests.